avoid pathological geospatial-index ordering
I'm seeing records condensed into a small geometry, giving a lack of
coverage, and the wrong overall impression of the point situation.

My theory is that postgres is returning whatever records which are
at-hand that satisfy the criteria, and there is somethign implementation
specific making geospatially near records more at hand.